Move sheep sheet data migration to proper function
