+++ module/ship-sheet.js
@@ -1,6 +1,5 @@
 import { diceRollType } from "./rolling/dice-rolling.js";
 import { RollModifier} from "./rolling/modifiers.js";
-import { migrateShipData } from "./shipDataMigration.js";
 
 Hooks.on('ready', () => {
     $(document).on('click', '.collateralDamageButton', async function (event) {
@@ -76,7 +75,6 @@ export class ExpanseShipSheet extends foundry.appv1.sheets.ActorSheet {
 
     async getData() {
         const sheetData = super.getData();
-        migrateShipData(this.actor);
 
         sheetData.dtypes = ["String", "Number", "Boolean"];
